chore(lint): migrate to ruff 0.16
Ruff 0.16 widened the default rule set far beyond the old E/F baseline
(bandit `S`, `BLE`, `B`, `C4`, `DTZ`, `PERF`, `SIM`, `I`, `UP`... now on by
default), taking `ruff check src/ tests/` from 0 to 654 findings on untouched
code. The pin at 0.15.9 kept preflight and CI in agreement; this bumps it and
does the cleanup in the same commit, as that pin's comment prescribed.

- 290 findings auto-fixed (`--fix`, no `--unsafe-fixes`): import sorting
  (I001), PEP 604 optionals (UP045), `datetime.UTC` (UP017), unused noqa
  (RUF100), deprecated imports (UP035).
- 13 fixed by hand: SIM115 context managers (cli.py daemon log fd, edit_ops
  x2), SIM117 nested `with` (test_resources x2), PIE810 startswith tuple,
  PERF402 list copy, PLW0602 needless `global`, PLW0211 staticmethod `cls`
  arg, RUF012 mutable class default (`_tools` -> tuple), F401, EXE002 (cli.py
  is a package module, not a script -- exec bit dropped), TRY004 suppressed
  with rationale (the loader's uniform RuntimeError contract is load-bearing).
- db_core: `except (sqlite3.OperationalError, AttributeError, Exception)`
  reduced to `except Exception` -- the narrow arms were already subsumed.

BLE001 / S110 / S112 (~275 sites) are ignored in [tool.ruff.lint] rather than
"fixed". They flag deliberate isolation boundaries -- an MCP server must not
let telemetry, embeddings, sqlite-vec, the watcher or a best-effort cleanup
kill the agent's session. Narrowing each to a concrete exception type would be
a behaviour change at every site with no defect behind it. Rationale is in the
config so the next reader does not have to re-derive it.

`ruff check src/ tests/` exits 0 on 0.16.0; 2073 passed, 2 skipped.

+++ src/token_savior/memory/budget.py
@@ -104,7 +104,7 @@ def format_session_budget_box(stats: dict[str, Any]) -> str:
     """Render get_session_budget_stats() as a 60-char status box."""
     pct = stats.get("pct_used", 0.0)
     bar_w = 40
-    filled = max(0, min(bar_w, int(round(pct / 100.0 * bar_w))))
+    filled = max(0, min(bar_w, round(pct / 100.0 * bar_w)))
     bar = "█" * filled + "·" * (bar_w - filled)
     sid = stats.get("session_id") or "—"
     project = stats.get("project_root") or "(none)"
